+++ LinearList/MinDistance.h
@@ -11,5 +11,5 @@
 @interface MinDistance : NSObject
 
 //求数值数组中大小最接近的两个元素的差
--(NSInteger)minDistanceOfArray:(NSInteger *)array withBegin:(NSInteger)p withEnd:(NSInteger)r;
+-(int)minDistanceOfArray:(int *)array withBegin:(int)p withEnd:(int)r;
 @end
