+++ LinearList/MinDistance.m
@@ -10,7 +10,7 @@
 
 @implementation MinDistance
 
--(NSInteger)minDistanceOfArray:(NSInteger *)array withBegin:(NSInteger)p withEnd:(NSInteger)r
+-(int)minDistanceOfArray:(int *)array withBegin:(int)p withEnd:(int)r
 {
     //除了快排没有找到更好的解法，待解决
     return 0;
