+++ Sort/Sort.m
@@ -12,34 +12,34 @@
 @implementation Sort
 
 #pragma mark - 快速排序
--(NSInteger *)quickSortOfArray:(NSInteger *)array withBegin:(NSInteger)p withEnd:(NSInteger)r
+-(int *)quickSortOfArray:(int *)array withBegin:(int)p withEnd:(int)r
 {
     if(p < r){
-        NSInteger q = [self partitionOfArray:array withBegin:p withEnd:r];
+        int q = [self partitionOfArray:array withBegin:p withEnd:r];
         [self quickSortOfArray:array withBegin:q+1 withEnd:r];
         [self quickSortOfArray:array withBegin:p withEnd:q-1];
     }
     return array;
 }
 
--(NSInteger)partitionOfArray:(NSInteger *)array withBegin:(NSInteger)p withEnd:(NSInteger)r
+-(int)partitionOfArray:(int *)array withBegin:(int)p withEnd:(int)r
 {
     //选取最右边元素作为主元
-    NSInteger mainElement = array[r];
-    NSInteger i = p - 1;//i初始指向开始的前一个位置
+    int mainElement = array[r];
+    int i = p - 1;//i初始指向开始的前一个位置
     //注意次处的指针j后移到r-1便停止，因为最后一个元素是主元
-    for (NSInteger j = p; j < r; j++) {
+    for (int j = p; j < r; j++) {
         if (array[j] <= mainElement) {
             //如果比主元小，就放到第i+1个位置，同时把该元素与j位置元素交换，也就是放到比x小的位置，
             //如果比主元大，指针j后移。i指针不变
             i++;
-            NSInteger temp = array[i];
+            int temp = array[i];
             array[i] = array[j];
             array[j] = temp;
         }
     }
     //最后把主元放到i+1位置，把该位置的数放到r位置
-    NSInteger temp = array[i+1];
+    int temp = array[i+1];
     array[i+1] = array[r];
     array[r] = temp;
     //返回主元所在的位置，这样主元前面的元素比他小，后面的元素比他大
@@ -48,21 +48,21 @@
 
 
 //随机选取主元的位置，并把该位置的元素放到r处，作为主元位置
--(NSInteger *)randomizedQuickSortOfArray:(NSInteger *)array withBegin:(NSInteger)p withEnd:(NSInteger)r
+-(int *)randomizedQuickSortOfArray:(int *)array withBegin:(int)p withEnd:(int)r
 {
     if(p < r){
-        NSInteger q = [self randomizedPartitionOfArray:array withBegin:p withEnd:r];
+        int q = [self randomizedPartitionOfArray:array withBegin:p withEnd:r];
         [self randomizedQuickSortOfArray:array withBegin:q+1 withEnd:r];
         [self randomizedQuickSortOfArray:array withBegin:p withEnd:q-1];
     }
     return array;
 }
 
--(NSInteger)randomizedPartitionOfArray:(NSInteger *)array withBegin:(NSInteger)p withEnd:(NSInteger)r
+-(int)randomizedPartitionOfArray:(int *)array withBegin:(int)p withEnd:(int)r
 {
     srand((unsigned) time(NULL));//以时间作为种子
-    NSInteger i = rand()%(r-p) + p;//随机生成一个位置，使该位置的元素作为主元，放到r位置
-    NSInteger temp = array[i];
+    int i = rand()%(r-p) + p;//随机生成一个位置，使该位置的元素作为主元，放到r位置
+    int temp = array[i];
     array[i] = array[r];
     array[r] = temp;
     return [self partitionOfArray:array withBegin:p withEnd:r];
@@ -70,14 +70,14 @@
 
 
 #pragma mark - 堆排序
--(Heap)heapSortOfArray:(NSInteger *)array withLength:(NSInteger)size
+-(Heap)heapSortOfArray:(int *)array withLength:(int)size
 {
     BinaryHeap *binaryHeap = [[BinaryHeap alloc] init];
-    //首先简历最大堆
+    //首先建立最大堆
     Heap heap =[binaryHeap buildMaxHeap:array withSize:size];
-    for(NSInteger i = size; i >= 2; i--){
+    for(int i = size; i >= 2; i--){
         //将堆的顶点与i位置的元素交换
-        NSInteger temp = heap->elements[1];
+        int temp = heap->elements[1];
         heap->elements[1] = heap->elements[i];
         heap->elements[i] = temp;
         heap->size --;
