+++ Tree/Tree.h
@@ -5,7 +5,7 @@
 //  Created by 浩 董 on 15/8/21.
 //  Copyright (c) 2015年 com.tencent. All rights reserved.
 //
-typedef NSInteger ElementType;
+typedef int ElementType;
 
 #ifndef Base_Tree_h
 #define Base_Tree_h
